chore(component): add alert dialog component for delete expense button

+++ src/app/(authenticated)/dashboard/expenses/_components/all-expenses/delete-expense-button.tsx
@@ -1,4 +1,15 @@
 import { deleteExpense } from "@/actions";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import { IconTrash } from "@irsyadadl/paranoid";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
@@ -33,13 +44,32 @@ export const DeleteExpenseButton = async ({ id }: { id: string }) => {
   };
 
   return (
-    <Button
-      variant="outline"
-      size="icon"
-      onClick={() => handleDelete(id)}
-      disabled={isPending}
-    >
-      {isPending ? "..." : <IconTrash className="h-4 w-4" />}
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button variant="outline" size="icon" disabled={isPending}>
+          <IconTrash className="h-4 w-4" />
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. Data will be permanently deleted from
+            the server.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <Button variant="destructive" asChild>
+            <AlertDialogAction
+              onClick={() => handleDelete(id)}
+              disabled={isPending}
+            >
+              Delete
+            </AlertDialogAction>
+          </Button>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 };
